Fix issues where `clickDisabled` is set to true too often (#435)
* Use minimum wait time before resetting clickDisabled
* Prevent disabling click during negligible mouse movement

When dragging is true, it has been noted that users may find themselves trying to click on an image and nothing happens. This is because without them knowing, `onMouseMove` has occurred with negligible movement to the user, but noticeable by the code which therefore disables clicking for the user. This is not optimal UX.

"10px" was chosen because this gives a user some leeway when they are trying to click something while accidentally barely moving the slide.